# agent-harness/src/marketing_os/entrypoints/api/app.py
"""FastAPI service exposing the Marketing OS graph over HTTP.

Endpoints:
  GET  /health
  POST /campaigns                       -> scaffold a campaign goal from the template
  GET  /campaigns/{slug}/gate?customer= -> Stage 0 gate report
  GET  /campaigns/{slug}/deliverables   -> list written deliverables
  POST /campaigns/{slug}/run            -> start a background run, return its run_id (202)
  GET  /runs                            -> list in-flight runs
  GET  /runs/{run_id}                   -> report a run's lifecycle status
  POST /runs/{run_id}/cancel            -> cancel an in-flight run
  GET  /runs/{run_id}/stream            -> attach to a run and tail its trace as SSE

Run with:  uvicorn marketing_os.entrypoints.api.app:app --reload
"""

from __future__ import annotations

import json
from collections.abc import AsyncIterator
from contextlib import asynccontextmanager
from dataclasses import asdict
from functools import lru_cache

from fastapi import FastAPI, HTTPException
from fastapi.responses import StreamingResponse
from pydantic import BaseModel

from marketing_os.adapters.documents import FilesystemDocumentStore
from marketing_os.adapters.observability import (
    configure_logging,
    configure_tracing,
    list_run_ids,
    new_run_id,
    read_events,
    tail_trace,
)
from marketing_os.config import Settings, load_settings
from marketing_os.entrypoints.env import load_env
from marketing_os.errors import GateError, MarketingOSError, RunConflictError
from marketing_os.governance import check_gate
from marketing_os.graph.registry import (
    CANCELLED,
    RUNNING,
    RunRegistry,
    read_run_status,
    resolve_trace_path,
)
from marketing_os.graph.runner import arun_campaign
from marketing_os.schemas import CampaignResult

load_env()


@asynccontextmanager
async def _lifespan(_: FastAPI) -> AsyncIterator[None]:
    """Configure logging and LangSmith tracing for the service lifetime.

    Args:
        _: The FastAPI application (unused).

    Yields:
        Control for the duration of the application's lifespan.
    """
    settings = get_settings()
    configure_logging(settings)
    configure_tracing(settings)
    yield


app = FastAPI(title="Marketing OS", version="0.2.0", lifespan=_lifespan)


def _http_error(exc: MarketingOSError) -> HTTPException:
    """Map a harness error to an HTTP error using the error's own presentation.

    The status code and structured payload come from the exception itself
    (``http_status`` and ``detail``), so the taxonomy is not re-spelled at each
    endpoint.

    Args:
        exc: The harness error to translate.

    Returns:
        The HTTP exception carrying the error's status and detail payload.
    """
    return HTTPException(exc.http_status, exc.detail or {"message": str(exc)})


@lru_cache(maxsize=1)
def get_settings() -> Settings:
    """Return the cached harness settings.

    Returns:
        The process-wide :class:`Settings` instance.
    """
    return load_settings()


@lru_cache(maxsize=1)
def get_document_store() -> FilesystemDocumentStore:
    """Return the process-wide document store tenant documents resolve through.

    Returns:
        The filesystem adapter rooted at the configured repo root (tests reset
        it with ``get_document_store.cache_clear()``, mirroring settings).
    """
    return FilesystemDocumentStore(get_settings().root)


@lru_cache(maxsize=1)
def get_registry() -> RunRegistry:
    """Return the process-wide registry of active background runs.

    The registry is a single in-memory instance for the process lifetime (tests
    reset it with ``get_registry.cache_clear()``, mirroring :func:`get_settings`).

    Returns:
        The shared :class:`RunRegistry`.
    """
    return RunRegistry()


class CreateCampaign(BaseModel):
    """Request body for scaffolding a campaign.

    Attributes:
        customer: The customer name.
        slug: The campaign slug; defaults to the customer name.
    """

    customer: str
    slug: str | None = None


class RunCampaign(BaseModel):
    """Request body for running a campaign.

    Attributes:
        customer: The customer name.
        stage: The single stage to run, or ``None`` for the full pipeline.
    """

    customer: str
    stage: str | None = None


@app.get("/health")
def health() -> dict[str, str]:
    """Report service health and the active provider and root.

    Returns:
        A status payload.
    """
    settings = get_settings()
    return {"status": "ok", "provider": settings.provider, "root": str(settings.root)}


@app.post("/campaigns")
def create_campaign(body: CreateCampaign) -> dict[str, object]:
    """Scaffold a campaign goal from the template and report the gate.

    Args:
        body: The create-campaign request.

    Returns:
        The slug, whether the goal was created, and the gate status.

    Raises:
        HTTPException: If the campaign-goal template is missing.
    """
    settings = get_settings()
    store = get_document_store()
    slug = body.slug or body.customer
    goal_document = f"campaigns/{slug}/goal.md"
    created = False
    if not store.exists(body.customer, goal_document):
        template = settings.templates_dir / "campaign-goal.md"
        if not template.is_file():
            raise HTTPException(500, "campaign-goal template missing")
        store.write(body.customer, goal_document, template.read_text(encoding="utf-8"))
        created = True
    report = check_gate(settings, body.customer, slug, store=store)
    return {
        "slug": slug,
        "customer": body.customer,
        "goal_created_from_template": created,
        "gate_ok": report.ok,
        "gate_issues": report.all_issues,
    }


@app.get("/campaigns/{slug}/gate")
def gate(slug: str, customer: str) -> dict[str, object]:
    """Return the Stage 0 gate report for a campaign.

    Args:
        slug: The campaign slug.
        customer: The customer name.

    Returns:
        The gate status and any issues.
    """
    settings = get_settings()
    report = check_gate(settings, customer, slug, store=get_document_store())
    return {"ok": report.ok, "issues": report.all_issues}


@app.get("/campaigns/{slug}/deliverables")
def deliverables(slug: str) -> dict[str, object]:
    """List the deliverable files written for a campaign.

    Args:
        slug: The campaign slug.

    Returns:
        The campaign slug and the list of written files.

    Raises:
        HTTPException: If the campaign directory does not exist.
    """
    settings = get_settings()
    campaign_dir = settings.campaigns_dir / slug
    if not campaign_dir.is_dir():
        raise HTTPException(404, f"No campaign '{slug}'")
    files = [
        {"name": f.name, "size_bytes": f.stat().st_size} for f in sorted(campaign_dir.glob("*.md"))
    ]
    return {"slug": slug, "files": files}


@app.post("/campaigns/{slug}/run", status_code=202)
async def run(slug: str, body: RunCampaign) -> dict[str, object]:
    """Start a detached background run and return its ``run_id`` immediately.

    The run is a first-class background job: it executes as an :class:`asyncio.Task`
    on the async graph path (ADR-0009), held in the process run registry keyed by
    slug. This endpoint no longer blocks on the pipeline — observe the run via
    :func:`get_run_status` (``GET /runs/{run_id}``) or the stream endpoint. The
    Stage 0 gate is checked synchronously so a misconfigured campaign fails fast
    rather than spawning a job that would immediately halt.

    Args:
        slug: The campaign slug.
        body: The run request.

    Returns:
        The new run's id, slug, stage, and initial ``running`` status.

    Raises:
        HTTPException: 409 if the gate failed or the slug already has an active run.
    """
    settings = get_settings()
    report = check_gate(settings, body.customer, slug, store=get_document_store())
    if not report.ok:
        raise _http_error(GateError("Stage 0 gate failed", missing=report.all_issues))
    run_id = new_run_id()

    async def launch() -> CampaignResult:
        """Execute the background run to completion on the async graph path.

        Returns:
            The structured campaign result.
        """
        return await arun_campaign(settings, body.customer, slug, stage=body.stage, run_id=run_id)

    try:
        get_registry().start(
            run_id=run_id,
            slug=slug,
            stage=body.stage,
            customer=body.customer,
            launch=launch,
        )
    except RunConflictError as exc:
        raise _http_error(exc) from exc
    return {"run_id": run_id, "slug": slug, "stage": body.stage, "status": RUNNING}


@app.get("/runs")
def list_active_runs() -> dict[str, object]:
    """List the runs currently in flight across all campaigns.

    Returns:
        The active runs, each with its ``run_id``, ``slug``, ``stage``, and
        ``customer``.
    """
    runs = [
        {"run_id": run.run_id, "slug": run.slug, "stage": run.stage, "customer": run.customer}
        for run in get_registry().active()
    ]
    return {"runs": runs}


@app.get("/runs/{run_id}")
def get_run_status(run_id: str) -> dict[str, object]:
    """Report a run's lifecycle status across its five terminal and live states.

    Resolves ``running`` from the live registry, or ``completed`` / ``failed`` /
    ``cancelled`` / ``interrupted`` from the run's JSONL trace.

    Args:
        run_id: The run id to query.

    Returns:
        The run's id, slug, stage, and status.

    Raises:
        HTTPException: 404 if the run id is unknown (no live run and no trace).
    """
    status = read_run_status(get_settings(), get_registry(), run_id)
    if status is None:
        raise HTTPException(404, f"No run '{run_id}'")
    return asdict(status)


@app.post("/runs/{run_id}/cancel")
async def cancel_run(run_id: str) -> dict[str, object]:
    """Cancel an in-flight run, aborting its in-flight LLM call.

    Cancelling the run's task lands a :class:`asyncio.CancelledError` inside the
    specialist's awaited LLM call (ADR-0009); the trace ends with a terminal
    ``run.summary outcome=cancelled`` event and the run leaves the live registry.

    Args:
        run_id: The id of the run to cancel.

    Returns:
        The cancelled run's id, slug, and ``cancelled`` status.

    Raises:
        HTTPException: 404 if no live run has that id (already finished or unknown).
    """
    cancelled = await get_registry().cancel(run_id)
    if cancelled is None:
        raise HTTPException(404, f"No active run '{run_id}'")
    return {"run_id": run_id, "slug": cancelled.slug, "status": CANCELLED}


@app.get("/campaigns/{slug}/runs")
def list_runs(slug: str) -> dict[str, object]:
    """List the run-log traces recorded for a campaign.

    Args:
        slug: The campaign slug.

    Returns:
        The campaign slug and the available run ids (newest first).
    """
    settings = get_settings()
    return {"slug": slug, "runs": list_run_ids(settings.logs_dir, slug)}


@app.get("/campaigns/{slug}/runs/{run_id}")
def get_run(slug: str, run_id: str) -> dict[str, object]:
    """Return the parsed JSONL trace for one run.

    Args:
        slug: The campaign slug.
        run_id: The run id (trace filename without extension).

    Returns:
        The campaign slug, run id, and the list of trace events.

    Raises:
        HTTPException: 404 if the trace does not exist.
    """
    settings = get_settings()
    path = settings.logs_dir / slug / f"{run_id}.jsonl"
    if not path.is_file():
        raise HTTPException(404, f"No run '{run_id}' for campaign '{slug}'")
    return {"slug": slug, "run_id": run_id, "events": read_events(path)}


@app.get("/runs/{run_id}/stream")
def stream_run(run_id: str) -> StreamingResponse:
    """Attach to an existing run and stream its progress as Server-Sent Events.

    Observing is split from starting (the run is already executing as a detached
    background job — see ``POST /run``): this endpoint does **not** launch a run, it
    tails the run's JSONL trace. A client attaching late is replayed the events
    already recorded from the top of the trace, then followed live until the terminal
    ``run.summary`` event, at which point the stream closes. Because it only reads the
    durable trace, any number of observers can attach to the same run concurrently,
    and a finished run replays and closes.

    Args:
        run_id: The id of the run to observe.

    Returns:
        A streaming response emitting one SSE ``data:`` frame per trace event.

    Raises:
        HTTPException: 404 if no live run and no trace exist for the run id.
    """
    settings = get_settings()
    registry = get_registry()
    trace_path = resolve_trace_path(settings, registry, run_id)
    if trace_path is None:
        raise HTTPException(404, f"No run '{run_id}'")

    async def event_source() -> AsyncIterator[str]:
        """Yield each trace event as an SSE ``data:`` frame.

        Yields:
            SSE-formatted event lines.
        """
        async for event in tail_trace(trace_path, is_live=lambda: registry.get(run_id) is not None):
            yield f"data: {json.dumps(event)}\n\n"

    return StreamingResponse(event_source(), media_type="text/event-stream")