# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: we-OS API
  version: 0.1.0
  contact:
    name: we-OS engineering
    url: https://github.com/ZAF07/we-os
  description: |
    The frozen contract between the we-OS frontend and the engine. The frontend
    codes against this document (served by a mock server) while the engine is
    built behind it; changes are deliberate and visible to both sides.

    ## Tenancy

    Identity is carried as a Bearer JWT issued by the managed IdP. The tenant —
    exactly one business — is derived from the **verified** claim on every
    request. **No operation accepts a tenant, business, or customer identity as
    a parameter.** A request without a valid token receives a 401
    `unauthenticated` error. A request whose token is valid but whose target
    resource belongs to a different tenant receives a 404 `not_found` — the
    refusal is indistinguishable from the resource not existing, so nothing
    leaks across the tenant boundary.

    ## Stages, phases, and lifecycle

    The API speaks **engine stages** (`research` … `asset-prompts`, in the
    mandatory order with the performance plan before the creative brief). Each
    stage also reports the operator **Phase** it belongs to (`Research`,
    `Strategy`, `Plan`, `Produce`) so the frontend renders its designed stepper
    without the engine adopting UI vocabulary. `Brief` is the campaign goal (a
    Stage 0 input), and `Approve` / `Publish` / `Measure` are campaign
...
    stages.

    ## Errors

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      summary: List the tenant's campaigns
      description: Everything in flight, with lifecycle status and stage progress separated.
      tags: [campaigns]
      responses:
        "200":
          description: Every campaign with its lifecycle status and stage progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
              example:
                campaigns:
                  - id: cmp_spring
                    name: Spring refill push
                    status: awaiting_approval
                    stage_progress:
                      completed: 1
                      total: 6
                      current_stage_key: brand-strategy
                    blocked_reason: Brand strategy is waiting for your approval.
                    updated_at: "2026-08-19T16:04:00Z"
                  - id: cmp_winter
                    name: Winter warmers
                    status: draft
                    stage_progress:
                      completed: 0
                      total: 6
                      current_stage_key: null
                    blocked_reason: null
                    updated_at: "2026-08-02T09:12:00Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      operationId: createCampaign
      summary: Create a campaign
      description: |
        The goal states one measurable business objective, a timeframe, the
        campaign budget (media/ad spend), the audience segment it targets, and
        all three KPI tiers. Channels are **not** part of the goal — we-OS
        chooses them in the performance plan.
      tags: [campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignCreate"
            example:
              name: Spring refill push
              objective: 120 refill subscriptions in 8 weeks
              timeframe:
                start_date: "2026-09-01"
                end_date: "2026-10-27"
              budget:
                amount: 4000
                currency: SGD
              audience_segment: Weekday regulars
              kpis:
                business: 120 refill subscriptions
                marketing: 2.5% landing-page conversion
                creative: 30% hook rate on launch video
      responses:
        "201":
          description: The created campaign, in `draft`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
              examples:
                draft:
                  $ref: "#/components/examples/CampaignDraft"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /campaigns/{campaignId}:
    parameters:
      - $ref: "#/components/parameters/CampaignId"
    get:
      operationId: getCampaign
      summary: Read a campaign
      description: The campaign, its goal fields, lifecycle status, and per-stage state.
      tags: [campaigns]
      responses:
        "200":
          description: The campaign, its goal, and every stage's state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
              examples:
                awaiting-approval:
                  $ref: "#/components/examples/CampaignAwaitingApproval"
                draft:
                  $ref: "#/components/examples/CampaignDraft"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /campaigns/{campaignId}/runs:
    parameters:
      - $ref: "#/components/parameters/CampaignId"
    get:
      operationId: listRuns
      summary: List a campaign's runs
      description: Run history, newest first — one entry per execution attempt.
      tags: [runs]
      responses:
        "200":
          description: The campaign's runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
              example:
                runs:
                  - run_id: run_01j8
                    campaign_id: cmp_spring
                    status: awaiting_approval
                    current_stage_key: brand-strategy
                    awaiting_approval_stage_key: brand-strategy
                    started_at: "2026-08-19T15:58:00Z"
                    finished_at: null
                  - run_id: run_01j2
                    campaign_id: cmp_spring
                    status: cancelled
                    current_stage_key: research
                    awaiting_approval_stage_key: null
                    started_at: "2026-08-18T11:20:00Z"
                    finished_at: "2026-08-18T11:24:00Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: startRun
      summary: Start a run
      description: |
        Refused with `gate_failed` (naming every missing Required field) when
        the Brand DNA or campaign goal is incomplete, with `run_conflict` when
        the campaign already has an active run, and with `quota_exhausted` when
        the tenant's allowance is spent. The run executes detached: close the
        tab and come back, or a deploy restarts the service, and the run is
        still there.
      tags: [runs]
      responses:
        "202":
          description: The run has started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
              example:
                run_id: run_01j8
                campaign_id: cmp_spring
                status: running
                current_stage_key: research
                awaiting_approval_stage_key: null
                started_at: "2026-08-19T15:58:00Z"
                finished_at: null
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "402":
          $ref: "#/components/responses/QuotaExhausted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The gate failed, or a run is already active for this campaign.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GateFailedError"
                  - $ref: "#/components/schemas/RunConflictError"
              examples:
                gate-failed:
                  summary: Brand DNA incomplete
                  value:
                    type: gate_failed
                    status: 409
                    message: "The DNA Gate failed: 2 Required fields are missing."
                    missing_fields:
                      - question_id: q_price_point
                        field: Price point
                        label: What do your main products or services cost?
                      - question_id: q_budget_range
                        field: Budget range
                        label: What can you spend on marketing in a typical month?
                run-conflict:
                  summary: A run is already active
                  value:
                    type: run_conflict
                    status: 409
                    message: Campaign 'cmp_spring' already has an active run.
                    active_run_id: run_01j8
  /runs/{runId}:
    get:
      operationId: getRun
      summary: Read a run's status
      description: Where the attempt has got to — live, halted at a gate, or terminal.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: The run and where it has got to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
              examples:
                running:
                  summary: Executing a stage
                  value:
                    run_id: run_01j8
                    campaign_id: cmp_spring
                    status: running
                    current_stage_key: research
                    awaiting_approval_stage_key: null
                    started_at: "2026-08-19T15:58:00Z"
                    finished_at: null
                awaiting-approval:
                  summary: Halted at an Approval Gate
                  value:
                    run_id: run_01j8
                    campaign_id: cmp_spring
                    status: awaiting_approval
                    current_stage_key: brand-strategy
                    awaiting_approval_stage_key: brand-strategy
                    started_at: "2026-08-19T15:58:00Z"
                    finished_at: null
                completed:
                  summary: Finished the pipeline
                  value:
                    run_id: run_01j8
                    campaign_id: cmp_spring
                    status: completed
                    current_stage_key: asset-prompts
                    awaiting_approval_stage_key: null
                    started_at: "2026-08-19T15:58:00Z"
                    finished_at: "2026-08-19T16:41:00Z"
                failed:
                  summary: Halted on an error the operator can act on
                  value:
                    run_id: run_01j8
                    campaign_id: cmp_spring
                    status: failed
                    current_stage_key: performance-plan
                    awaiting_approval_stage_key: null
                    started_at: "2026-08-19T15:58:00Z"
                    finished_at: "2026-08-19T16:12:00Z"
                    failure:
                      type: guardrail
                      message: The performance plan could not meet the quality bar within the revision budget.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /runs/{runId}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel a run
      description: |
        Cancelling abandons the attempt: the next run of the campaign starts
        clean from stage 1 rather than resuming.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: The cancelled run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
              example:
                run_id: run_01j8
                campaign_id: cmp_spring
                status: cancelled
                current_stage_key: research
                awaiting_approval_stage_key: null
                started_at: "2026-08-19T15:58:00Z"
                finished_at: "2026-08-19T16:02:00Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT from the managed IdP. The BFF verifies it for rendering and
        forwards it; the engine verifies it independently and derives the
        tenant from the verified claim.
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      description: The campaign id (always scoped to the caller's tenant).
      schema:
        type: string
      example: cmp_spring
    RunId:
      name: runId
      in: path
      required: true
      description: The run id (always scoped to the caller's tenant).
      schema:
        type: string
      example: run_01j8
...
  responses:
    Unauthenticated:
      description: The request carries no valid identity.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: unauthenticated
            status: 401
            message: Sign in to continue.
    NotFound:
      description: |
        No such resource in the caller's tenant. Cross-tenant access resolves
        here too — a foreign id is indistinguishable from a missing one.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: not_found
            status: 404
            message: No such campaign.
    ValidationFailed:
      description: The request body is malformed or incomplete.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: validation
            status: 422
            message: "kpis.creative is required: every campaign defines all three KPI tiers."
    QuotaExhausted:
      description: The tenant's allowance is spent; billable work is refused before it starts.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/QuotaExhaustedError"
          example:
            type: quota_exhausted
            status: 402
            message: Your allowance is used up. Work resumes when it renews.
            used: 100
            allowance: 100
  schemas:
...
    Timeframe:
      type: object
      required: [start_date, end_date]
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
        currency:
          type: string
          description: ISO 4217 code.
...
    CampaignCreate:
      type: object
      required: [name, objective, timeframe, budget, audience_segment, kpis]
      properties:
        name:
          type: string
        objective:
          type: string
          description: One measurable business objective.
        timeframe:
          $ref: "#/components/schemas/Timeframe"
        budget:
          $ref: "#/components/schemas/Money"
        audience_segment:
          type: string
          description: One of the audience segments named in the Brand DNA.
        kpis:
          $ref: "#/components/schemas/KpiTiers"
    CampaignStatus:
      type: string
      description: Campaign lifecycle — a separate axis from stage progress.
      enum: [draft, running, awaiting_approval, approved, published, measuring, archived]
...
    Campaign:
      type: object
      required: [id, name, objective, timeframe, budget, audience_segment, kpis, status, stages, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        objective:
          type: string
        timeframe:
          $ref: "#/components/schemas/Timeframe"
        budget:
          $ref: "#/components/schemas/Money"
        audience_segment:
          type: string
        kpis:
          $ref: "#/components/schemas/KpiTiers"
        status:
          $ref: "#/components/schemas/CampaignStatus"
        stages:
          type: array
          description: All six engine stages in pipeline order, each with its Phase, state, and approval policy.
          items:
            $ref: "#/components/schemas/Stage"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CampaignSummary:
      type: object
      required: [id, name, status, stage_progress, blocked_reason, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/CampaignStatus"
        stage_progress:
          $ref: "#/components/schemas/StageProgress"
        blocked_reason:
          type: string
          nullable: true
          description: What is blocking the campaign, in operator language, or null.
        updated_at:
          type: string
          format: date-time
    CampaignList:
      type: object
      required: [campaigns]
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/CampaignSummary"
    RunStatus:
      type: string
      enum: [running, awaiting_approval, completed, failed, cancelled, interrupted]
...
    Run:
      type: object
      required: [run_id, campaign_id, status, current_stage_key, awaiting_approval_stage_key, started_at, finished_at]
      properties:
        run_id:
          type: string
        campaign_id:
          type: string
        status:
          $ref: "#/components/schemas/RunStatus"
        current_stage_key:
          $ref: "#/components/schemas/NullableStageKey"
        awaiting_approval_stage_key:
          description: Set exactly when the run is halted at an Approval Gate.
          allOf:
            - $ref: "#/components/schemas/NullableStageKey"
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        failure:
          $ref: "#/components/schemas/RunFailure"
    RunList:
      type: object
      required: [runs]
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"
...
    Error:
      type: object
      required: [type, status, message]
      properties:
        type:
          type: string
          enum:
            - unauthenticated
            - not_found
            - validation
            - gate_failed
            - quota_exhausted
            - run_conflict
            - stage_not_awaiting_approval
            - revision_limit_reached
            - internal
        status:
          type: integer
          description: The HTTP status this error carries.
        message:
          type: string
    GateFailedError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [missing_fields]
          properties:
            missing_fields:
              type: array
              description: Every missing Required field, named exactly.
              items:
                $ref: "#/components/schemas/MissingField"
...
    RunConflictError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [active_run_id]
          properties:
            active_run_id:
              type: string
  examples:
    CampaignDraft:
      summary: A freshly created campaign
      value:
        id: cmp_spring
        name: Spring refill push
        objective: 120 refill subscriptions in 8 weeks
        timeframe:
          start_date: "2026-09-01"
          end_date: "2026-10-27"
        budget:
          amount: 4000
          currency: SGD
        audience_segment: Weekday regulars
        kpis:
          business: 120 refill subscriptions
          marketing: 2.5% landing-page conversion
          creative: 30% hook rate on launch video
        status: draft
        stages:
          - {key: research, phase: Research, state: pending, approval_policy: auto}
          - {key: brand-strategy, phase: Strategy, state: pending, approval_policy: human}
          - {key: campaign-strategy, phase: Strategy, state: pending, approval_policy: human}
          - {key: performance-plan, phase: Plan, state: pending, approval_policy: human}
          - {key: creative-brief, phase: Produce, state: pending, approval_policy: human}
          - {key: asset-prompts, phase: Produce, state: pending, approval_policy: human}
        created_at: "2026-08-19T15:40:00Z"
        updated_at: "2026-08-19T15:40:00Z"
    CampaignAwaitingApproval:
      summary: Halted at the brand-strategy Approval Gate
      value:
        id: cmp_spring
        name: Spring refill push
        objective: 120 refill subscriptions in 8 weeks
        timeframe:
          start_date: "2026-09-01"
          end_date: "2026-10-27"
        budget:
          amount: 4000
          currency: SGD
        audience_segment: Weekday regulars
        kpis:
          business: 120 refill subscriptions
          marketing: 2.5% landing-page conversion
          creative: 30% hook rate on launch video
        status: awaiting_approval
        stages:
          - {key: research, phase: Research, state: completed, approval_policy: auto}
          - {key: brand-strategy, phase: Strategy, state: awaiting_approval, approval_policy: human}
          - {key: campaign-strategy, phase: Strategy, state: pending, approval_policy: human}
          - {key: performance-plan, phase: Plan, state: pending, approval_policy: human}
          - {key: creative-brief, phase: Produce, state: pending, approval_policy: human}
          - {key: asset-prompts, phase: Produce, state: pending, approval_policy: human}
        created_at: "2026-08-19T15:40:00Z"
        updated_at: "2026-08-19T16:04:00Z"
...